feat(as-3509): update azure pipeline config
