+++ packages/common/src/utils.js
@@ -1,21 +1,4 @@
-/**
- * Utility methods
- */
-
-const toArray = (candidate) => {
-  const array = [];
-  if (candidate) {
-    if (Array.isArray(candidate)) {
-      array.push(...candidate);
-    } else {
-      array.push(candidate);
-    }
-  }
-  return array;
-};
-
 module.exports = {
-  toArray,
   /**
    * Promise Timeout
    *
