+++ packages/web-app/package.json
@@ -13,7 +13,8 @@
     "build": "npm run build:sass && npm run build:app",
     "build:app": "cp -Rf ./src ./dist",
     "build:sass": "sass src/sass/main.scss:src/public/stylesheets/main.css --style=compressed --load-path=./ --no-source-map",
-    "start": "node dist/server.js"
+    "start": "node dist/server.js",
+    "documentation": "jest --bail --coverage --coverageReporters=cobertura --reporters=default --reporters=jest-junit"
   },
   "lint-staged": {
     "./**/*.js": "npm run lint:fix",
@@ -21,6 +22,7 @@
   },
   "dependencies": {
     "@pins/common": "file:../common",
+    "@testing-library/jest-dom": "^5.14.1",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.5",
     "dotenv": "^10.0.0",
@@ -28,6 +30,9 @@
     "express-pino-logger": "^6.0.0",
     "express-validator": "^6.12.1",
     "govuk-frontend": "^3.13.0",
+    "jest": "^27.0.6",
+    "jest-fetch-mock": "^3.0.3",
+    "jest-junit": "^13.0.0",
     "lusca": "^1.7.0",
     "node-fetch": "^2.6.2",
     "nunjucks": "^3.2.3",
@@ -39,7 +44,6 @@
     "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
     "babel-eslint": "^10.1.0",
     "eslint": "^7.32.0",
     "eslint-config-airbnb-base": "^14.2.1",
@@ -47,8 +51,6 @@
     "eslint-plugin-import": "^2.24.1",
     "eslint-plugin-jest": "^24.4.0",
     "eslint-plugin-prettier": "^3.4.1",
-    "jest": "^27.0.6",
-    "jest-fetch-mock": "^3.0.3",
     "nodemon": "^2.0.12",
     "pino-pretty": "^5.1.3",
     "prettier": "^2.3.2"
